+++ src/parsePosts.js
@@ -11,6 +11,13 @@ import templates from './utils/templates.js';
 const fm = require('front-matter');
 const rss = require('rss');
 
+const configFile = Bun.file(join(import.meta.dir, '../config.json'));
+
+if (!await configFile.exists()) {
+    console.log('You need to make a config.json to use bun-blog. You can start by copying or renaming config.json.sample to config.json');
+    process.exit();
+}
+
 const mdPosts = await walk(join(import.meta.dir, '../posts'));
 
 if (await fs.exists(join(import.meta.dir, '../public/posts'))) await fs.rm(join(import.meta.dir, '../public/posts'), { recursive: true, force: true });
